+++ extensions/shell/CHANGELOG.md
@@ -1,16 +1,28 @@
 # shell Changelog
 
+## [Search History] 2024-01-12
+
+- Added actions to delete/restore your git history
+- works with zsh or bash history
+- search through all of your history
+- create a ray.so link
+- copy the command to your clipboard
+
 ## [Fixes] 2023-01-02
+
 Fix an issue when the use Terminal/iTerm option is turned-on and command run with argument by moving execution logic to useEffect.
 
 ## [Add Warp Support] - 2022-11-18
-This version adds the option to run commands in Warp. 
+
+This version adds the option to run commands in Warp.
 Checkout the preferences and select your favorite console application in the dropdown to get going.
 
 ## [Update] - 2022-09-10
+
 Open new iTerm tab if iTerm is running and chooses Execute in iTerm.app.
 
 ## [iTerm fixes] - 2022-06-30
+
 fix an issue, where commands did not run, when iTerm was not open.
 
 ## [Initial Version] - 2022-06-21